Scheduler: move GetAvoidPodsFromNodeAnnotations to component-helpers
