+++ staging/src/k8s.io/component-helpers/scheduling/corev1/helpers.go
@@ -17,6 +17,8 @@ limitations under the License.
 package corev1
 
 import (
+	"encoding/json"
+
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/component-helpers/scheduling/corev1/nodeaffinity"
 )
@@ -42,4 +44,17 @@ func MatchNodeSelectorTerms(
 		return false, nil
 	}
 	return nodeaffinity.NewLazyErrorNodeSelector(nodeSelector).Match(node)
+}
+
+// GetAvoidPodsFromNodeAnnotations scans the list of annotations and
+// returns the pods that needs to be avoided for this node from scheduling
+func GetAvoidPodsFromNodeAnnotations(annotations map[string]string) (v1.AvoidPods, error) {
+	var avoidPods v1.AvoidPods
+	if len(annotations) > 0 && annotations[v1.PreferAvoidPodsAnnotationKey] != "" {
+		err := json.Unmarshal([]byte(annotations[v1.PreferAvoidPodsAnnotationKey]), &avoidPods)
+		if err != nil {
+			return avoidPods, err
+		}
+	}
+	return avoidPods, nil
 }
